requirements.txt
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu126
python-dotenv
requests
fastapi
uvicorn
openai

langchain==0.3.26 
langchain-openai 
langchain-pinecone 
langchain-community 
langchain-text-splitters 
langchain-core 
pinecone-client
pytesseract 
pillow
PyMuPDF